wip: PT mostly working except for PG -> PT assets

+++ code/projects/offline_renderer/pt_scene.cpp
@@ -194,7 +194,7 @@ bool Scene::Load( const std::string& filename )
     LOG( "Building BVH..." );
     auto bvhTime = Time::GetTimePoint();
     bvh.Build( shapes );
-    float bvhBuildTime = Time::GetDuration( bvhTime ) / 1000.0f;
+    float bvhBuildTime = (float)Time::GetDuration( bvhTime ) / 1000.0f;
     LOG( "BVH build time: %.2f seconds", bvhBuildTime );
 
     return true;
